cv24_luganda/requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
torch>=2.1.0
torchaudio>=2.1.0
datasets>=2.19.0
huggingface_hub>=0.23.0

# ── Audio ─────────────────────────────────────────────────────────────────────
soundfile>=0.12.1

# ── Data ──────────────────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
pyarrow>=15.0.0

# ── HTTP / download ───────────────────────────────────────────────────────────
requests>=2.31.0

# ── CLI / UX ──────────────────────────────────────────────────────────────────
click>=8.1.7
rich>=13.7.0
python-dotenv>=1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
